move options decorator under pieces subdirectory
