+++ src/lib/decorators/pieces/options.decorator.ts
@@ -1,4 +1,4 @@
-import { PieceOptions } from '../structures/pieces/piece';
+import { PieceOptions } from '../../structures/pieces/piece';
 
 function ApplyOptions<T extends PieceOptions>(options: T) {
   return function <E extends { new (...args: any[]): {} }>(target: E): any {
